// src/app/game-model/pattern-type.ts
export enum PatternType {
  Cell = 'Cell',
  Glider = 'Glider',
  Block = 'Block',
  Beehive = 'Beehive',
  Loaf = 'Loaf',
  Boat = 'Boat',
  Blinker = 'Blinker',
  Toad = 'Toad',
  Beacon = 'Beacon',
  Pulsar = 'Pulsar',
  LWSS = 'LWSS',
  MWSS = 'MWSS',
  GosperGun = 'Gosper Gun',
  RPentomino = 'R-pentomino'
}